* scale translate done 0.5

+++ src/webgl/shaders/vertext-shader.vert
@@ -4,7 +4,7 @@ attribute vec2 aTextureCoord;
 uniform mat4 uModelViewMatrix;
 uniform mat4 uProjectionMatrix;
 
-varying highp vec2 vTextureCoord;
+varying mediump vec2 vTextureCoord;
 
 void main(void) {
     gl_Position = uProjectionMatrix * uModelViewMatrix * aVertexPosition;
